requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx~=0.25.0
sqlalchemy==2.0.23
alembic==1.12.1
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
click==8.1.7
rich==13.7.0
tabulate==0.9.0
python-dateutil==2.8.2
psutil==5.9.6
structlog
sentry-sdk[fastapi]
pybreaker
passlib
pytest-cov # Added for test coverage
